requirements.txt
python-telegram-bot==20.6
google-generativeai==0.3.2
supabase==1.2.0
python-dotenv==1.0.0
apscheduler==3.10.4